requirements.txt
numpy~=1.22.3
PyOpenGL~=3.1.6
PyOpenGL-accelerate~=3.1.5
glfw~=2.5.1
